+++ src/main/java/model/card/Ace.java
@@ -1,4 +1,4 @@
-package model;
+package model.card;
 
 public class Ace extends Card {
