+++ src/main/java/model/card/Denomination.java
@@ -1,4 +1,4 @@
-package model;
+package model.card;
 
 public enum Denomination {
     ACE("A"),
