+++ src/main/java/model/casino/Deck.java
@@ -1,8 +1,14 @@
-package model;
+package model.casino;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import model.card.Ace;
+import model.card.Card;
+import model.card.Denomination;
+import model.card.Suit;
+import model.participants.Participants;
+import model.participants.Player;
 
 public class Deck {
 
@@ -28,7 +34,7 @@ public class Deck {
     public void distributeCards(Participants participants) {
         shuffle();
         drawTwoCardFromDeck(participants.getDealer());
-        for (Player player : participants.getPlayers().getPlayers()) {
+        for (Player player : participants.getPlayerGroup().getPlayers()) {
             drawTwoCardFromDeck(player);
         }
     }
