+++ src/main/java/view/InputView.java
@@ -4,8 +4,8 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Scanner;
 import java.util.regex.Pattern;
-import model.Player;
-import model.Players;
+import model.participants.Player;
+import model.participants.PlayerGroup;
 
 public class InputView {
     private static final String PLAYER_NAMES_INPUT_REGEX = "^([가-힣a-zA-Z]+)(,\s*[가-힣a-zA-Z]+)*$";
@@ -17,9 +17,9 @@ public class InputView {
 
     private static final Scanner sc = new Scanner(System.in);
 
-    public static Map<Player, Integer> inputBettingPrice(Players players) {
+    public static Map<Player, Integer> inputBettingPrice(PlayerGroup playerGroup) {
         Map<Player, Integer> bettingPrices = new HashMap<>();
-        for (Player player: players.getPlayers()) {
+        for (Player player : playerGroup.getPlayers()) {
             StringBuilder stringBuilder = new StringBuilder();
             stringBuilder.append(player.getName())
                     .append("의 배팅 금액은?");
