+++ src/test/java/model/DeckTest.java
@@ -4,6 +4,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.ArrayList;
 import java.util.List;
+import model.card.Card;
+import model.card.Denomination;
+import model.card.Suit;
+import model.casino.Deck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
