+++ src/test/java/model/HandTest.java
@@ -1,9 +1,13 @@
 package model;
 
-import static model.Denomination.ACE;
-import static model.Suit.CLUB;
+import static model.card.Denomination.ACE;
+import static model.card.Suit.CLUB;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import model.card.Ace;
+import model.card.Card;
+import model.card.Denomination;
+import model.card.Hand;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
