+++ src/test/java/model/PlayerTest.java
@@ -2,6 +2,11 @@ package model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import model.card.Ace;
+import model.card.Card;
+import model.card.Denomination;
+import model.card.Suit;
+import model.participants.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
